requirements.txt
certifi==2020.12.5
chardet==4.0.0
Django==2.2.20
django-cors-headers==3.7.0
djangorestframework==3.12.4
fake-useragent==0.1.11
idna==2.10
numpy==1.20.2
pandas==1.2.4
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
selenium==3.141.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.4
